Finish lesson 06-cascade-fix
Interesting to see that there are several different ways that you could
do this one. I didn't take the exact approach that they took. In terms
of how clean the code looks I think I prefer the way that they recommend
but I think that my method was a quicker way of fixing it.