Finished the MDN Countdown
Took a few goes to get this right - made a few infinite loops but
understand the reasoning for this now.

Interesting to note that I did this using a for loop while MDN did this
using a while loop. Don't think it makes a difference to the outcome
at all but good to know that there are different ways to do things.